[Style] #14 UIFont extension 위치 옮김
