+++ Bingha/Bingha/Extensions/UIFontExtension.swift
@@ -1,0 +1,17 @@
+//
+//  UIFont.swift
+//  Bingha
+//
+//  Created by Terry Koo on 2022/07/20.
+//
+
+import UIKit
+
+extension UIFont {
+    class func rounded(ofSize size: CGFloat, weight: UIFont.Weight) -> UIFont {
+        let systemFont = UIFont.systemFont(ofSize: size, weight: weight)
+
+        guard #available(iOS 13.0, *), let descriptor = systemFont.fontDescriptor.withDesign(.rounded) else { return systemFont }
+        return UIFont(descriptor: descriptor, size: size)
+    }
+}
