+++ Bingha/Bingha/Views/Iceberg/IcebergViewController.swift
@@ -53,13 +53,3 @@ class IcebergViewController: UIViewController {
         view.addSubview(circularProgressBarView)
    }
 }
-
-
-extension UIFont {
-    class func rounded(ofSize size: CGFloat, weight: UIFont.Weight) -> UIFont {
-        let systemFont = UIFont.systemFont(ofSize: size, weight: weight)
-
-        guard #available(iOS 13.0, *), let descriptor = systemFont.fontDescriptor.withDesign(.rounded) else { return systemFont }
-        return UIFont(descriptor: descriptor, size: size)
-    }
-}
